Added music, improved movment
look at my code for movement in obj_another_player and slow enemy and
try to emulate it as your code while good, was really wasteful, it was
good to get a start though. I still have to add health and other stuff,
but i figure you could do that once we get to that point.