Adds Enemy, Minor Object-Oriented Changes
Added the bit enemy we discussed. It works fine except for lacking
collision (like every other enemy, it's just more notable with the
collision damage ones).

Saved a bunch of code that should apply to every enemy as .txt files so
we can just load them in the future. Tried putting them in obj_enemy so
they'd all inherit, but it didn't work.