Overhaul for Upgrade System
Changed everything to have 100x the damage and health it previously did.
No current visible changes, but allows for more granular damage upgrades
(advancing from 100 to 150 instead of 1 to 2 or 1 to 1.5) (In my
experience, decimals are harder to 'get' instinctively, so this was why
I went with 100x)

Also changed the player to have a set of globals that dictate their
actions, and for functions that care about the player's stats (bullet
collisions with enemies for the damage variable, for example) to
reference these instead of hardcoding.